refactor: modulariza telas de perfil e acessibilidade
